+++ elfi/methods/parameter_inference.py
@@ -22,6 +22,7 @@ from elfi.methods.utils import (GMDistribution, ModelPrior, arr2d_to_batch,
                                 batch_to_arr2d, ceil_to_batch_size, weighted_var)
 from elfi.model.elfi_model import ComputationContext, ElfiModel, NodeReference
 from elfi.utils import is_array
+from elfi.visualization.visualization import progress_bar
 
 logger = logging.getLogger(__name__)
 
@@ -234,10 +235,17 @@ class ParameterInference:
         """
         raise NotImplementedError
 
-    def infer(self, *args, vis=None, **kwargs):
+    def infer(self, *args, vis=None, bar=True, **kwargs):
         """Set the objective and start the iterate loop until the inference is finished.
 
         See the other arguments from the `set_objective` method.
+
+        Parameters
+        ----------
+        vis : dict, optional
+            Plotting options. More info in self.plot_state method
+        bar : bool, optional
+            Flag to remove (False) or keep (True) the progress bar from/in output.
 
         Returns
         -------
@@ -248,10 +256,18 @@ class ParameterInference:
 
         self.set_objective(*args, **kwargs)
 
+        if bar:
+            progress_bar(0, self._objective_n_batches, prefix='Progress:',
+                         suffix='Complete', length=50)
+
         while not self.finished:
             self.iterate()
             if vis:
                 self.plot_state(interactive=True, **vis_opt)
+
+            if bar:
+                progress_bar(self.state['n_batches'], self._objective_n_batches,
+                             prefix='Progress:', suffix='Complete', length=50)
 
         self.batches.cancel_pending()
         if vis:
@@ -296,14 +312,12 @@ class ParameterInference:
         return self._objective_n_batches <= self.state['n_batches']
 
     def _allow_submit(self, batch_index):
-        return self.max_parallel_batches > self.batches.num_pending and \
-            self._has_batches_to_submit and \
-            (not self.batches.has_ready())
+        return (self.max_parallel_batches > self.batches.num_pending and
+                self._has_batches_to_submit and (not self.batches.has_ready()))
 
     @property
     def _has_batches_to_submit(self):
-        return self._objective_n_batches > \
-            self.state['n_batches'] + self.batches.num_pending
+        return self._objective_n_batches > self.state['n_batches'] + self.batches.num_pending
 
     @property
     def _objective_n_batches(self):
@@ -387,7 +401,9 @@ class Sampler(ParameterInference):
         result : Sample
 
         """
-        return self.infer(n_samples, *args, **kwargs)
+        bar = kwargs.pop('bar', True)
+
+        return self.infer(n_samples, *args, bar=bar, **kwargs)
 
     def _extract_result_kwargs(self):
         kwargs = super(Sampler, self)._extract_result_kwargs()
@@ -837,8 +853,7 @@ class BayesianOptimization(ParameterInference):
         super(BayesianOptimization, self).__init__(
             model, output_names, batch_size=batch_size, **kwargs)
 
-        target_model = target_model or \
-            GPyRegression(self.model.parameter_names, bounds=bounds)
+        target_model = target_model or GPyRegression(self.model.parameter_names, bounds=bounds)
 
         self.target_name = target_name
         self.target_model = target_model
@@ -851,12 +866,11 @@ class BayesianOptimization(ParameterInference):
             self.target_model.update(params, precomputed[target_name])
 
         self.batches_per_acquisition = batches_per_acquisition or self.max_parallel_batches
-        self.acquisition_method = acquisition_method or \
-            LCBSC(self.target_model,
-                  prior=ModelPrior(self.model),
-                  noise_var=acq_noise_var,
-                  exploration_rate=exploration_rate,
-                  seed=self.seed)
+        self.acquisition_method = acquisition_method or LCBSC(self.target_model,
+                                                              prior=ModelPrior(self.model),
+                                                              noise_var=acq_noise_var,
+                                                              exploration_rate=exploration_rate,
+                                                              seed=self.seed)
 
         self.n_initial_evidence = n_initial
         self.n_precomputed_evidence = n_precomputed
@@ -1136,7 +1150,7 @@ class BOLFI(BayesianOptimization):
 
     """
 
-    def fit(self, n_evidence, threshold=None):
+    def fit(self, n_evidence, threshold=None, bar=True):
         """Fit the surrogate model.
 
         Generates a regression model for the discrepancy given the parameters.
@@ -1145,8 +1159,12 @@ class BOLFI(BayesianOptimization):
 
         Parameters
         ----------
+        n_evidence : int, required
+            Number of evidence for fitting
         threshold : float, optional
             Discrepancy threshold for creating the posterior (log with log discrepancy).
+        bar : bool, optional
+            Flag to remove (False) the progress bar from output.
 
         """
         logger.info("BOLFI: Fitting the surrogate model...")
@@ -1155,7 +1173,7 @@ class BOLFI(BayesianOptimization):
             raise ValueError(
                 'You must specify the number of evidence (n_evidence) for the fitting')
 
-        self.infer(n_evidence)
+        self.infer(n_evidence, bar=bar)
         return self.extract_posterior(threshold)
 
     def extract_posterior(self, threshold=None):