+++ tests/conftest.py
@@ -4,6 +4,7 @@ import time
 
 import numpy as np
 import pytest
+import scipy.stats as ss
 
 import elfi
 import elfi.clients.ipyparallel as eipp
@@ -25,8 +26,6 @@ def pytest_addoption(parser):
         action="store",
         default="all",
         help="perform the tests for the specified client (default all)")
-
-    parser.addoption("--skipslow", action="store_true", help="skip slow tests")
 
 
 """Functional fixtures"""
@@ -189,15 +188,27 @@ def sleeper(sec, batch_size, random_state):
     return secs
 
 
+@pytest.fixture()
+def multivariate_model(request):
+    ndim = request.param
+
+    def fun(x, batch_size, random_state):
+        return np.sum(x, keepdims=True, axis=1)
+
+    m = elfi.ElfiModel()
+    elfi.Prior(ss.multivariate_normal, [0]*ndim, model=m, name='t1')
+    elfi.Simulator(fun, m['t1'], observed=np.array([[0]]), model=m, name='sim')
+    elfi.Distance('euclidean', m['sim'], model=m, name='d')
+    return m
+
+
 def no_op(data):
     return data
 
 
 @pytest.fixture()
 def sleep_model(request):
-    """The true param will be half of the given sleep time
-
-    """
+    """The true param will be half of the given sleep time."""
     ub_sec = request.param or .5
     m = elfi.ElfiModel()
     elfi.Constant(ub_sec, model=m, name='ub')
@@ -207,6 +218,20 @@ def sleep_model(request):
     elfi.Distance('euclidean', m['summary'], model=m, name='d')
 
     m.observed['slept'] = ub_sec / 2
+    return m
+
+
+def rowsummer(x, batch_size, random_state):
+    return np.sum(x, keepdims=True, axis=1)
+
+
+@pytest.fixture()
+def multivariate_model(request):
+    ndim = request.param
+    m = elfi.ElfiModel()
+    elfi.Prior(ss.multivariate_normal, [0]*ndim, model=m, name='t1')
+    elfi.Simulator(rowsummer, m['t1'], observed=np.array([[0]]), model=m, name='sim')
+    elfi.Distance('euclidean', m['sim'], model=m, name='d')
     return m